Implementación de la seguridad

+++ src/main/java/upc/edu/pe/happypaws/serviceimplements/RolServiceImplement.java
@@ -29,7 +29,7 @@ public class RolServiceImplement implements IRolService {
     }
 
     @Override
-    public void delete(int rolId) {
+    public void delete(Long rolId) {
         rolRepository.deleteById(rolId);
     }
